working through unsigned long long to qint64 and code cleanup

+++ digdeeperdialog.cpp
@@ -4,7 +4,7 @@
 // Copyright 2015 Pasquale J. Rinaldi, Jr.
 // Distrubted under the terms of the GNU General Public License version 2
 
-DigDeeperDialog::DigDeeperDialog(QWidget *parent, unsigned long long curcheckcount, unsigned long long curlistcount) :
+DigDeeperDialog::DigDeeperDialog(QWidget *parent, qint64 curcheckcount, qint64 curlistcount) :
     QDialog(parent),
     ui(new Ui::DigDeeperDialog)
 {
